requirements.txt
aiohttp==3.6.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2018.8.24
chardet==3.0.4
colorzero==1.1
cryptography==2.6.1
distro-info==0.21
entrypoints==0.3
future==0.18.2
gpiozero==1.5.1
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
multidict==4.7.3
PiHole-api==2.0
pycrypto==2.6.1
PyGObject==3.30.4
PySocks==1.7.1
python-apt==1.8.4
python-telegram-bot==12.2.0
pyxdg==0.25
requests==2.21.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
ssh-import-id==5.7
tornado==6.0.3
unattended-upgrades==0.1
urllib3==1.24.1
yarl==1.4.2
